Implement getting crc32
If "useCRC32" is set, then it will use the CRCpp library to parse it from the file being added to the playlist.  This should either be compatable with the crc32's in RA's databases, or may break things.  Getting the CRC32 directly from RA's databases will be implemented soon.
also tweaked the systems.json, don't actually need the . before extensions.
also update switch-nanogui.