Implement Explorer classes
uses Filesystem to traverse folders.
Explorer is a List that handles filling itself with the list of folders/files at a given path
FileItem and FolderItem are simple classes to make things pretty, FileItem will change color depending on if the extention is known or not.  This will possibly change later.
updated settings.json to reflect what Filesystem wants to see (no "/" at the end)